backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
motor>=3.3.0
pymongo>=4.5.0
python-dotenv>=0.19.0,<0.20.0
beautifulsoup4>=4.9.3,<4.10.0
requests>=2.26.0,<2.27.0
google-generative-ai>=0.2.0
tavily-python>=0.1.9
langchain>=0.0.200
celery>=5.2.7
youtube-dl>=2021.12.17
weasyprint>=54.3
pydantic>=1.8.0,<2.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
